move the insertDiagnostico method to MySQL.java
