+++ src/main/java/com/meddetectai/main/Paciente.java
@@ -77,21 +77,6 @@ public class Paciente {
         return currentPaciente;
     }
 
-    public static void insertDiagnostico(Diagnostico diagnostico, String cpfPaciente) {
-        try (Connection conn = DriverManager.getConnection(MySQL.getUrl(), MySQL.getUser(), MySQL.getPassword())) {
-            String sql = "INSERT INTO Diagnostico (cpf_paciente, tipo, imagem, resultado) VALUES (?, ?, ?, ?) " +
-                         "ON DUPLICATE KEY UPDATE tipo = VALUES(tipo), imagem = VALUES(imagem), resultado = VALUES(resultado)";
-            PreparedStatement ps = conn.prepareStatement(sql);
-            ps.setString(1, cpfPaciente);
-            ps.setString(2, diagnostico.getTipo().name());
-            ps.setBytes(3, diagnostico.getImagem());
-            ps.setString(4, diagnostico.getResultado());
-            ps.executeUpdate();
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }
-    }
-
     @Override
     public String toString() {
         return this.getNome();
